refactor: move regression_evaluate to infer_true_shower_parameters
