+++ src/infer_true_shower_parameters/baseline/decision_tree_regression.py
@@ -6,7 +6,7 @@ from sklearn.tree import DecisionTreeRegressor
 from src.infer_true_shower_parameters import NUM_TRUE_SHOWER
 import pandas as pd
 import pickle
-from src.common.regression_evaluate import mse, r2
+from src.infer_true_shower_parameters.regression_evaluate import mse, r2
 from pathlib import Path
 
 class DecisionTreeRegression(nn.Module):
